+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 * X and Y range selection items:
   * Added support for item title in parameters data set (`RangeShapeParam`)
   * This concerns the `XRangeSelection` and `YRangeSelection` items
+* New annotated X and Y range selection items:
+  * Added `AnnotatedXRangeSelection` and `AnnotatedYRangeSelection` items
+  * These items provide X and Y range selection with an annotation label
+  * They can be created using `make.annotated_xrange` and `make.annotated_yrange` functions
 
 🧹 API cleanup: removed deprecated update methods (use `update_item` instead)